WIP - start to incorporate chunk deletion code from WorldBorder
Experiment - The goal is to strip this down to just the chunk deletion
parts and use it for chunk deletion of islands. Then when chunks need to
be created they will use the chunk generation of the appropriate Game
Mode Addon. This will also necessitate island boundaries to be on chunk
boundaries.